Add test for YAMLReadyDictDatasourceConfigSerializer
